+++ src/lib/Server/Common/Request.ts
@@ -16,7 +16,7 @@
 
 import { IRawRequest } from '../../Common';
 
-export interface IRequest<T extends any[] = any[]> extends IRawRequest<T> {
+export interface IRequest<T extends any = any> extends IRawRequest<T> {
 
     srt: number;
 }